Feat : 현재위치 가져오는 hook 변동
